Modified handleFailureMessage to handle String-only response.
In cases where the server's response is, for instance
"Unauthorized", this method currently ignores the message
as it is not a JSONObject or a JSONArray. A String /is/ a valid JSON
value though, legitimately returned by JSONTokener. I have therefore
added an extra else-if branch to call onFailure(Throwable, String)
in the parent class for such situations.