Check if entity's stream is really compressed.
One of our clients has a server returning 404 errors with
Content-Encoding=gzip, but the content itself is not compressed, causing
'new GZIPInputStream' in  InflatingEntity#getContent to throw an
IOException.

When using FileAsyncHttpResponseHandler, for some reason we couldn't
find out, after some requests with the wrong 404 response,
AsyncHttpClient was getting stuck and after some time it raised
ConnectionPoolTimeoutException. This wasn't happening when the
BaseJsonHttpResponseHandler was being used.